LabelNameTexture使用固定可重用Bitmap，防止重复创建Bitmap，减少相册内滚动界面时的内存分配，进而减少界面卡顿。修复因ThumbnailLoader.Uploader上传已回收Texture而导致SlotView显示出错的BUG。

+++ app/src/main/java/com/gk969/UltimateImgSpider/StaticValue.java
@@ -64,7 +64,7 @@ public class StaticValue {
     public final static String EXTRA_URL_TO_OPEN = "urlToOpen";
 
     public static int getThumbnailCacheSize() {
-        return (int) (MemoryInfo.getTotalMemInMb() / 2048 + 1) * 64;
+        return (MemoryInfo.getTotalMemInMb() / 2048 + 1) * 64;
     }
 
     public static int getSpiderDownloaderThreadNum() {
